feat: CoreDataStorage 싱글톤으로 변경
